requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-ReCaptcha==0.4.2
Werkzeug==3.0.1
SQLAlchemy==2.0.23
alembic==1.13.1
email-validator==2.1.0.post1
python-dotenv==1.0.0
Pillow==10.1.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
pyinstaller==6.3.0
pandas==2.1.1
reportlab==3.5.64
XlsxWriter==3.1.9
requests==2.31.0
bcrypt==4.1.2
PyJWT==2.8.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
mysqlclient==2.2.1
redis==5.0.1
Celery==5.3.6
WTForms==3.1.1
Babel==2.14.0
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
webbrowser==0.10.2